Added moving average windowing to ADC
The ADC readings have quite a lot jitter (~1%). This change smooths
out the readings over a moving average.

This change also introduces an offset floor level for zero luminosity
to sidestep the issue of not being able to truly turn off the LED due
to the ADC readings not being fixed at zero.